Emails: Move In-Depth Comparison components under new folder
