+++ client/my-sites/email/controller.js
@@ -7,7 +7,7 @@ import TitanControlPanelRedirect from 'calypso/my-sites/email/email-management/t
 import TitanManageMailboxes from 'calypso/my-sites/email/email-management/titan-manage-mailboxes';
 import TitanManagementIframe from 'calypso/my-sites/email/email-management/titan-management-iframe';
 import EmailProvidersComparison from 'calypso/my-sites/email/email-providers-comparison';
-import EmailProvidersInDepthComparison from 'calypso/my-sites/email/email-providers-in-depth-comparison';
+import EmailProvidersInDepthComparison from 'calypso/my-sites/email/email-providers-comparison/in-depth';
 import EmailProvidersStackedComparison from 'calypso/my-sites/email/email-providers-stacked-comparison';
 import { castIntervalLength } from 'calypso/my-sites/email/email-providers-stacked-comparison/provider-cards/utils';
 import GSuiteAddUsers from 'calypso/my-sites/email/gsuite-add-users';
